+++ framework/core/Hooks/Hookable.php
@@ -1,0 +1,17 @@
+<?php
+/**
+ * Hook interface
+ *
+ * @package Brocooly
+ * @subpackage Brocket
+ * @since 1.8.7
+ */
+
+declare(strict_types=1);
+
+namespace Brocooly\Hooks;
+
+interface Hookable
+{
+	public function init();
+}